+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add `Colors` filter with `color_multiply!` and alias `*`
 - `bitmap_text` accepts multiline strings and supports colorization
 - `bitmap_text` supports left, center and right alignment
+- Add `BitmapText` filter for overlaying text onto images
 
 ## [0.2.0] - 2025-07-21
 - Ruby Sixel encoder now sets pixel aspect ratio metadata to display correctly in Windows Terminal